web/requirements.txt
Django==1.11.17
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.7
uWSGI==2.0.17.1
celery==4.2.1
flower==0.9.2
redis==3.0.1
requests==2.21.0
